refactor(tool-installer): move functions to organized dir
