+++ packages/@aglabo-actions/tool-installer/src/utils/__tests__/prepareInstallDirectory.spec.ts
