+++ packages/@aglabo-actions/tool-installer/src/utils/prepareInstallDirectory.ts
